rename oage.ts to pages.ts
