+++ src/blog/BlogPreview.tsx
@@ -3,7 +3,7 @@ import { QueryResponseInitial, useQuery } from "@sanity/react-loader";
 import { Suspense } from "react";
 
 import { BlogPage, Post } from "studio/lib/interfaces/pages";
-import { BLOG_PAGE_QUERY } from "studio/lib/queries/page";
+import { BLOG_PAGE_QUERY } from "studio/lib/queries/pages";
 
 import { Blog } from "./Blog";
