+++ src/components/sections/imageSection/ImageSectionComponentPreview.tsx
@@ -4,7 +4,7 @@ import { Suspense } from "react";
 
 import { PreviewProps } from "src/types/preview";
 import { ImageSection, PageBuilder } from "studio/lib/interfaces/pages";
-import { PAGE_QUERY } from "studio/lib/queries/page";
+import { PAGE_QUERY } from "studio/lib/queries/pages";
 
 import ImageSectionComponent from "./ImageSectionComponent";
