+++ studio/lib/queries/siteMap.ts
@@ -1,5 +1,6 @@
 import { groq } from "next-sanity";
 
+//Site Map
 export const DOCUMENTS_WITH_SLUG_QUERY = groq`
   *[defined(slug)]
 `;
