+++ studio/lib/queries/specialPages.ts
@@ -1,8 +1,10 @@
 import { groq } from "next-sanity";
 
-export const CUSTOMER_CASES_PAGE_QUERY = groq`
-  *[_type == "customerCasesPage" && slug.current == $slug][0]`;
-
+//Compensations
 export const COMPENSATIONS_PAGE_QUERY = groq`
   *[_type == "compensations" && slug.current == $slug][0]
 `;
+
+//Customer Cases
+export const CUSTOMER_CASES_PAGE_QUERY = groq`
+  *[_type == "customerCasesPage" && slug.current == $slug][0]`;
